fix some warnings

+++ HALBirdRecord/Model/HALActivityManager.m
@@ -47,7 +47,7 @@
     if (!self.activityList) {
         [self loadActivityList];
     }
-    return self.activityList.count;
+    return (int)self.activityList.count;
 }
 
 - (HALActivity *)activityWithIndex:(int)index
@@ -125,7 +125,7 @@
         activity.title = kHALDummyActivityName;
         for (int j=0; j < birdRecordCount; j++) {
             NSInteger birdID = arc4random() % 800;
-            HALBirdRecord *record = [HALBirdRecord birdRecordWithBirdID:birdID];
+            HALBirdRecord *record = [HALBirdRecord birdRecordWithBirdID:(int)birdID];
             int lat = (arc4random() % 25) + 20;
             int lng = (arc4random() % 30) + 122;
             record.coordinate = CLLocationCoordinate2DMake(lat, lng);
